refactor: move unique constaint to separate file
